[WFLY-19160] Refactor all multinode EJB over HTTP tests into one
