+++ testsuite/integration/multinode/src/test/java/org/jboss/as/test/multinode/ejb/http/EjbOverHttpTestCase.java
@@ -5,6 +5,7 @@
 
 package org.jboss.as.test.multinode.ejb.http;
 
+import jakarta.ejb.NoSuchEJBException;
 import org.jboss.arquillian.container.test.api.Deployer;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.OperateOnDeployment;
@@ -29,19 +30,16 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import javax.naming.Context;
 import javax.naming.InitialContext;
 import java.net.SocketPermission;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.Hashtable;
 
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.COMPOSITE;
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.OP;
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.OP_ADDR;
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.STEPS;
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.SUBSYSTEM;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.*;
 import static org.jboss.as.test.shared.PermissionUtils.createFilePermission;
 import static org.jboss.as.test.shared.PermissionUtils.createPermissionsXmlAsset;
-
 
 /**
  * @author <a href="mailto:tadamski@redhat.com">Tomasz Adamski</a>
@@ -52,6 +50,13 @@ public class EjbOverHttpTestCase {
     private static final Logger log = Logger.getLogger(EjbOverHttpTestCase.class);
     public static final String ARCHIVE_NAME_SERVER = "ejboverhttp-test-server";
     public static final String ARCHIVE_NAME_CLIENT = "ejboverhttp-test-client";
+
+    public static final String ARCHIVE_NAME_PROVIDER_URL_CLIENT = "ejboverhttp-test-providerurl-client";
+
+
+    public static final String ARCHIVE_NAME_DESCRIPTOR_URL_CLIENT = "ejboverhttp-test-descriptor-client";
+
+    public static final String ARCHIVE_NAME_WRONG_CREDENTIALS_CLIENT = "ejboverhttp-test-wrong-credentials-client";
     public static final int NO_EJB_RETURN_CODE = -1;
     private static final int serverPort = 8180;
 
@@ -100,25 +105,14 @@ public class EjbOverHttpTestCase {
 
     @Deployment(name = "server", managed = false)
     @TargetsContainer("multinode-server")
-    public static Archive<?> deployment0() {
-        JavaArchive jar = createJar(ARCHIVE_NAME_SERVER);
-        return jar;
+    public static Archive<?> serverDeployment() {
+        return createJar(ARCHIVE_NAME_SERVER);
+
     }
 
     @Deployment(name = "client")
     @TargetsContainer("multinode-client")
-    public static Archive<?> deployment1() {
-        JavaArchive clientJar = createClientJar();
-        return clientJar;
-    }
-
-    private static JavaArchive createJar(String archiveName) {
-        JavaArchive jar = ShrinkWrap.create(JavaArchive.class, archiveName + ".jar");
-        jar.addClasses(StatelessBean.class, StatelessLocal.class, StatelessRemote.class);
-        return jar;
-    }
-
-    private static JavaArchive createClientJar() {
+    public static Archive<?> clientDeployment() {
         JavaArchive jar = createJar(EjbOverHttpTestCase.ARCHIVE_NAME_CLIENT);
         jar.addClasses(EjbOverHttpTestCase.class);
         jar.addAsManifestResource("META-INF/jboss-ejb-client-profile.xml", "jboss-ejb-client.xml")
@@ -133,31 +127,195 @@ public class EjbOverHttpTestCase {
         return jar;
     }
 
+    @Deployment(name = "providerurl-client")
+    @TargetsContainer("multinode-client")
+    public static Archive<?> deploymentProviderUrlClient() {
+        JavaArchive jar = createJar(EjbOverHttpTestCase.ARCHIVE_NAME_PROVIDER_URL_CLIENT);
+        jar.addClasses(EjbOverHttpTestCase.class);
+        jar.addAsManifestResource("META-INF/jboss-ejb-client-profile.xml", "jboss-ejb-client.xml")
+                .addAsManifestResource("ejb-http-wildfly-config.xml", "wildfly-config.xml")
+                .addAsManifestResource(
+                        createPermissionsXmlAsset(
+                                createFilePermission("read,write,delete",
+                                        "jbossas.multinode.client", Arrays.asList("standalone", "data", "ejb-xa-recovery", "-")),
+                                new SocketPermission(TestSuiteEnvironment.formatPossibleIpv6Address(System.getProperty("node0")) + ":" + serverPort,
+                                        "connect,resolve")),
+                        "permissions.xml");
+        return jar;
+    }
+
+    @Deployment(name = "descriptor-client")
+    @TargetsContainer("multinode-client")
+    public static Archive<?> deploymentDescriptorClient() {
+        JavaArchive jar = createJar(EjbOverHttpTestCase.ARCHIVE_NAME_DESCRIPTOR_URL_CLIENT);
+        jar.addClasses(EjbOverHttpTestCase.class);
+        jar.addAsManifestResource("META-INF/jboss-ejb-client-http-connections.xml", "jboss-ejb-client.xml")
+                .addAsManifestResource("ejb-http-wildfly-config.xml", "wildfly-config.xml")
+                .addAsManifestResource(
+                        createPermissionsXmlAsset(
+                                createFilePermission("read,write,delete",
+                                        "jbossas.multinode.client", Arrays.asList("standalone", "data", "ejb-xa-recovery", "-")),
+                                new SocketPermission(TestSuiteEnvironment.formatPossibleIpv6Address(System.getProperty("node0")) + ":" + serverPort,
+                                        "connect,resolve")),
+                        "permissions.xml");
+        return jar;
+    }
+
+    @Deployment(name = "wrong-credentials-client")
+    @TargetsContainer("multinode-client")
+    public static Archive<?> deploymentWrongCredentialsClient() {
+        JavaArchive jar = createJar(EjbOverHttpTestCase.ARCHIVE_NAME_WRONG_CREDENTIALS_CLIENT);
+        jar.addClasses(EjbOverHttpTestCase.class);
+        jar.addAsManifestResource("META-INF/jboss-ejb-client-profile.xml", "jboss-ejb-client.xml")
+                .addAsManifestResource("ejb-http-wildfly-config-wrong.xml", "wildfly-config.xml")
+                .addAsManifestResource(
+                        createPermissionsXmlAsset(
+                                createFilePermission("read,write,delete",
+                                        "jbossas.multinode.client", Arrays.asList("standalone", "data", "ejb-xa-recovery", "-")),
+                                new SocketPermission(TestSuiteEnvironment.formatPossibleIpv6Address(System.getProperty("node0")) + ":" + serverPort,
+                                        "connect,resolve")),
+                        "permissions.xml");
+        return jar;
+    }
+
+    private static JavaArchive createJar(String archiveName) {
+        JavaArchive jar = ShrinkWrap.create(JavaArchive.class, archiveName + ".jar");
+        jar.addClasses(StatelessBean.class, StatelessLocal.class, StatelessRemote.class);
+        return jar;
+    }
+
     @Test
     @OperateOnDeployment("client")
     public void testBasicInvocation(@ArquillianResource InitialContext ctx) throws Exception {
-        deployer.deploy("server");
+        try {
+            deployer.deploy("server");
 
-        StatelessRemote bean = (StatelessRemote) ctx.lookup("java:module/" + StatelessBean.class.getSimpleName() + "!"
-                + StatelessRemote.class.getName());
-        Assert.assertNotNull(bean);
+            StatelessRemote bean = (StatelessRemote) ctx.lookup("java:module/" + StatelessBean.class.getSimpleName() + "!"
+                    + StatelessRemote.class.getName());
+            Assert.assertNotNull(bean);
 
-        // initial discovery
-        int methodCount = bean.remoteCall();
-        Assert.assertEquals(1, methodCount);
+            // initial discovery
+            int methodCount = bean.remoteCall();
+            Assert.assertEquals(1, methodCount);
 
-        deployer.undeploy("server");
+            deployer.undeploy("server");
 
-        //  failed discovery after undeploying server deployment
-        int returnValue = bean.remoteCall();
-        Assert.assertEquals(NO_EJB_RETURN_CODE, returnValue);
+            //  failed discovery after undeploying server deployment
+            int returnValue = bean.remoteCall();
+            Assert.assertEquals(NO_EJB_RETURN_CODE, returnValue);
 
-        deployer.deploy("server");
+            deployer.deploy("server");
 
-        // rediscovery after redeployment
-        methodCount = bean.remoteCall();
-        Assert.assertEquals(1, methodCount);
+            // rediscovery after redeployment
+            methodCount = bean.remoteCall();
+            Assert.assertEquals(1, methodCount);
+        } finally {
+            try {
+                deployer.undeploy("server");
+            } catch (Exception ignored) {
+            }
+        }
 
-        deployer.undeploy("server");
+
+    }
+
+    @Test
+    @OperateOnDeployment("providerurl-client")
+    public void testProviderUrlInvocation() throws Exception {
+        try {
+            deployer.deploy("server");
+
+            Hashtable<String, String> table = new Hashtable<>();
+            table.put(Context.INITIAL_CONTEXT_FACTORY, "org.wildfly.naming.client.WildFlyInitialContextFactory");
+            table.put(Context.PROVIDER_URL, "http://localhost:8180/wildfly-services");
+
+            InitialContext ctx = new InitialContext(table);
+
+            StatelessRemote bean = (StatelessRemote) ctx.lookup("java:ejboverhttp-test-server/" + StatelessBean.class.getSimpleName() + "!"
+                    + StatelessRemote.class.getName());
+            Assert.assertNotNull(bean);
+
+            // initial discovery
+            int methodCount = bean.remoteCall();
+            Assert.assertEquals(1, methodCount);
+
+            deployer.undeploy("server");
+
+            //  failed discovery after undeploying server deployment
+            boolean noSuchEjbException = false;
+            try {
+                int returnValue = bean.remoteCall();
+                Assert.assertEquals(NO_EJB_RETURN_CODE, returnValue);
+            } catch (NoSuchEJBException e) {
+                noSuchEjbException = true;
+            }
+            Assert.assertTrue(noSuchEjbException);
+
+            deployer.deploy("server");
+
+            // rediscovery after redeployment
+            methodCount = bean.remoteCall();
+            Assert.assertEquals(1, methodCount);
+        } finally {
+            try {
+                deployer.undeploy("server");
+            } catch (Exception ignored) {
+            }
+        }
+    }
+
+    @Test
+    @OperateOnDeployment("descriptor-client")
+    public void testDescriptorInvocation(@ArquillianResource InitialContext ctx) throws Exception {
+        try {
+            deployer.deploy("server");
+
+            StatelessRemote bean = (StatelessRemote) ctx.lookup("java:module/" + StatelessBean.class.getSimpleName() + "!"
+                    + StatelessRemote.class.getName());
+            Assert.assertNotNull(bean);
+
+            // initial discovery
+            int methodCount = bean.remoteCall();
+            Assert.assertEquals(1, methodCount);
+
+            deployer.undeploy("server");
+
+            //  failed discovery after undeploying server deployment
+            int returnValue = bean.remoteCall();
+            Assert.assertEquals(NO_EJB_RETURN_CODE, returnValue);
+
+            deployer.deploy("server");
+
+            // rediscovery after redeployment
+            methodCount = bean.remoteCall();
+            Assert.assertEquals(1, methodCount);
+        } finally {
+            try {
+                deployer.undeploy("server");
+            } catch (Exception ignored) {
+            }
+        }
+    }
+
+    @Test
+    @OperateOnDeployment("wrong-credentials-client")
+    public void testBasicInvocationWithWrongCredentials(@ArquillianResource InitialContext ctx) throws Exception {
+        try {
+            StatelessRemote bean = (StatelessRemote) ctx.lookup("java:module/" + StatelessBean.class.getSimpleName() + "!"
+                    + StatelessRemote.class.getName());
+            Assert.assertNotNull(bean);
+
+            try {
+                int methodCount = bean.remoteCall();
+                Assert.assertEquals(EjbOverHttpTestCase.NO_EJB_RETURN_CODE, methodCount);
+            } catch (javax.naming.AuthenticationException e) {
+                // expected
+            }
+        } finally {
+            try {
+                deployer.undeploy("server");
+            } catch (Exception ignored) {
+            }
+        }
     }
 }
